# be/src/main/resources/openapi/apispec.yml
openapi: 3.0.3
info:
  title: Employee and EmployeeDetail API
  description: API for managing Employees and EmployeeDetails
  version: 1.0.0
servers:
  - url: /api
    description: API Server
paths:
...
  /employees/{employeeId}:
    get:
      summary: Get an employee by ID
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          description: ID of the employee to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
    put:
      summary: Update an employee by ID
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          description: ID of the employee to update
          schema:
            type: integer
      requestBody:
        description: Updated employee object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
    delete:
      summary: Delete an employee by ID
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          description: ID of the employee to delete
          schema:
            type: integer
      responses:
        '204':
          description: Employee deleted
        '404':
          description: Employee not found
...
components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          description: Employee ID
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        employeeDetail:
          $ref: '#/components/schemas/EmployeeDetail'
    EmployeeDetail:
      type: object
      properties:
        id:
          type: integer
          description: EmployeeDetail ID
        department:
          type: string
        rank:
          type: integer
        salary:
          type: number
          format: float
        employee:
          $ref: '#/components/schemas/Employee'